Moved AppResources into app module, merged app_state.rs and app.rs
